+++ src/backend/access/rmgrdesc/gistdesc.c
@@ -25,18 +25,18 @@ out_gistxlogPageUpdate(StringInfo buf, gistxlogPageUpdate *xlrec)
 static void
 out_gistxlogPageReuse(StringInfo buf, gistxlogPageReuse *xlrec)
 {
-	appendStringInfo(buf, "rel %u/%u/%u; blk %u; snapshotConflictHorizon %llu, isCatalogRel %c",
+	appendStringInfo(buf, "rel %u/%u/%u; blk %u; snapshotConflictHorizon %" PRIu64 ", isCatalogRel %c",
 					 xlrec->locator.spcOid, xlrec->locator.dbOid,
 					 xlrec->locator.relNumber, xlrec->block,
-					 (unsigned long long) XidFromFullTransactionId(xlrec->snapshotConflictHorizon),
+					 XidFromFullTransactionId(xlrec->snapshotConflictHorizon),
 					 xlrec->isCatalogRel ? 'T' : 'F');
 }
 
 static void
 out_gistxlogDelete(StringInfo buf, gistxlogDelete *xlrec)
 {
-	appendStringInfo(buf, "delete: snapshotConflictHorizon %llu, nitems: %u, isCatalogRel %c",
-					 (unsigned long long) xlrec->snapshotConflictHorizon,
+	appendStringInfo(buf, "delete: snapshotConflictHorizon %" PRIu64 ", nitems: %u, isCatalogRel %c",
+					 xlrec->snapshotConflictHorizon,
 					 xlrec->ntodelete, xlrec->isCatalogRel ? 'T' : 'F');
 }
 
@@ -50,8 +50,8 @@ out_gistxlogPageSplit(StringInfo buf, gistxlogPageSplit *xlrec)
 static void
 out_gistxlogPageDelete(StringInfo buf, gistxlogPageDelete *xlrec)
 {
-	appendStringInfo(buf, "deleteXid %llu; downlink %u",
-					 (unsigned long long) XidFromFullTransactionId(xlrec->deleteXid),
+	appendStringInfo(buf, "deleteXid %" PRIu64 "; downlink %u",
+					 XidFromFullTransactionId(xlrec->deleteXid),
 					 xlrec->downlinkOffset);
 }
