+++ src/backend/access/rmgrdesc/xactdesc.c
@@ -306,7 +306,7 @@ xact_desc_subxacts(StringInfo buf, int nsubxacts, TransactionId *subxacts)
 	{
 		appendStringInfoString(buf, "; subxacts:");
 		for (i = 0; i < nsubxacts; i++)
-			appendStringInfo(buf, " %llu", (unsigned long long) subxacts[i]);
+			appendStringInfo(buf, " %" PRIu64, subxacts[i]);
 	}
 }
 
@@ -341,7 +341,7 @@ xact_desc_commit(StringInfo buf, uint8 info, xl_xact_commit *xlrec, RepOriginId 
 
 	/* If this is a prepared xact, show the xid of the original xact */
 	if (TransactionIdIsValid(parsed.twophase_xid))
-		appendStringInfo(buf, "%llu: ", (unsigned long long) parsed.twophase_xid);
+		appendStringInfo(buf, "%" PRIu64 ": ", parsed.twophase_xid);
 
 	appendStringInfoString(buf, timestamptz_to_str(xlrec->xact_time));
 
@@ -377,7 +377,7 @@ xact_desc_abort(StringInfo buf, uint8 info, xl_xact_abort *xlrec, RepOriginId or
 
 	/* If this is a prepared xact, show the xid of the original xact */
 	if (TransactionIdIsValid(parsed.twophase_xid))
-		appendStringInfo(buf, "%llu: ", (unsigned long long) parsed.twophase_xid);
+		appendStringInfo(buf, "%" PRIu64 ": ", parsed.twophase_xid);
 
 	appendStringInfoString(buf, timestamptz_to_str(xlrec->xact_time));
 
@@ -434,7 +434,7 @@ xact_desc_assignment(StringInfo buf, xl_xact_assignment *xlrec)
 	appendStringInfoString(buf, "subxacts:");
 
 	for (i = 0; i < xlrec->nsubxacts; i++)
-		appendStringInfo(buf, " %llu", (unsigned long long) xlrec->xsub[i]);
+		appendStringInfo(buf, " %" PRIu64, xlrec->xsub[i]);
 }
 
 void
@@ -473,7 +473,7 @@ xact_desc(StringInfo buf, XLogReaderState *record)
 		 * interested in the top-level xid that issued the record and which
 		 * xids are being reported here.
 		 */
-		appendStringInfo(buf, "xtop %llu: ", (unsigned long long) xlrec->xtop);
+		appendStringInfo(buf, "xtop %" PRIu64 ": ", xlrec->xtop);
 		xact_desc_assignment(buf, xlrec);
 	}
 	else if (info == XLOG_XACT_INVALIDATIONS)
