+++ src/backend/access/transam/transam.c
@@ -158,8 +158,8 @@ TransactionIdDidCommit(TransactionId transactionId)
 		parentXid = SubTransGetParent(transactionId);
 		if (!TransactionIdIsValid(parentXid))
 		{
-			elog(WARNING, "no pg_subtrans entry for subcommitted XID %llu",
-				 (unsigned long long) transactionId);
+			elog(WARNING, "no pg_subtrans entry for subcommitted XID %" PRIu64,
+				 transactionId);
 			return false;
 		}
 		return TransactionIdDidCommit(parentXid);
@@ -213,8 +213,8 @@ TransactionIdDidAbort(TransactionId transactionId)
 		if (!TransactionIdIsValid(parentXid))
 		{
 			/* see notes in TransactionIdDidCommit */
-			elog(WARNING, "no pg_subtrans entry for subcommitted XID %llu",
-				 (unsigned long long) transactionId);
+			elog(WARNING, "no pg_subtrans entry for subcommitted XID %" PRIu64,
+				 transactionId);
 			return true;
 		}
 		return TransactionIdDidAbort(parentXid);
