+++ src/backend/replication/logical/reorderbuffer.c
@@ -2655,8 +2655,8 @@ ReorderBufferProcessTXN(ReorderBuffer *rb, ReorderBufferTXN *txn,
 
 		/* this is just a sanity check against bad output plugin behaviour */
 		if (GetCurrentTransactionIdIfAny() != InvalidTransactionId)
-			elog(ERROR, "output plugin used XID %llu",
-				 (unsigned long long) GetCurrentTransactionId());
+			elog(ERROR, "output plugin used XID %" PRIu64,
+				 GetCurrentTransactionId());
 
 		/*
 		 * Remember the command ID and snapshot for the next set of changes in
@@ -3139,8 +3139,7 @@ ReorderBufferAbortOld(ReorderBuffer *rb, TransactionId oldestRunningXid)
 
 		if (TransactionIdPrecedes(txn->xid, oldestRunningXid))
 		{
-			elog(DEBUG2, "aborting old transaction %llu",
-				 (unsigned long long) txn->xid);
+			elog(DEBUG2, "aborting old transaction %" PRIu64, txn->xid);
 
 			/* Notify the remote node about the crash/immediate restart. */
 			if (rbtxn_is_streamed(txn))
@@ -3970,8 +3969,8 @@ ReorderBufferSerializeTXN(ReorderBuffer *rb, ReorderBufferTXN *txn)
 	Size		spilled = 0;
 	Size		size = txn->size;
 
-	elog(DEBUG2, "spill %u changes in XID %llu to disk",
-		 (uint32) txn->nentries_mem, (unsigned long long) txn->xid);
+	elog(DEBUG2, "spill %u changes in XID %" PRIu64 " to disk",
+		 (uint32) txn->nentries_mem, txn->xid);
 
 	/* do the same to all child TXs */
 	dlist_foreach(subtxn_i, &txn->subtxns)
@@ -4250,8 +4249,8 @@ ReorderBufferSerializeChange(ReorderBuffer *rb, ReorderBufferTXN *txn,
 		errno = save_errno ? save_errno : ENOSPC;
 		ereport(ERROR,
 				(errcode_for_file_access(),
-				 errmsg("could not write to data file for XID %llu: %m",
-						(unsigned long long) txn->xid)));
+				 errmsg("could not write to data file for XID %" PRIu64 ": %m",
+						txn->xid)));
 	}
 	pgstat_report_wait_end();
 
@@ -4894,10 +4893,10 @@ ReorderBufferSerializedPath(char *path, ReplicationSlot *slot, TransactionId xid
 
 	XLogSegNoOffsetToRecPtr(segno, 0, wal_segment_size, recptr);
 
-	snprintf(path, MAXPGPATH, "%s/%s/xid-%llu-lsn-%X-%X.spill",
+	snprintf(path, MAXPGPATH, "%s/%s/xid-%" PRIu64 "-lsn-%X-%X.spill",
 			 PG_REPLSLOT_DIR,
 			 NameStr(MyReplicationSlot->data.name),
-			 (unsigned long long) xid, LSN_FORMAT_ARGS(recptr));
+			 xid, LSN_FORMAT_ARGS(recptr));
 }
 
 /*
@@ -5520,8 +5519,8 @@ UpdateLogicalMappings(HTAB *tuplecid_data, Oid relid, Snapshot snapshot)
 	{
 		RewriteMappingFile *f = (RewriteMappingFile *) lfirst(file);
 
-		elog(DEBUG1, "applying mapping: \"%s\" in %llu", f->fname,
-			 (unsigned long long) snapshot->subxip[0]);
+		elog(DEBUG1, "applying mapping: \"%s\" in %" PRIu64, f->fname,
+			 snapshot->subxip[0]);
 		ApplyLogicalMappingFile(tuplecid_data, relid, f->fname);
 		pfree(f);
 	}
