+++ src/backend/utils/misc/help_config.c
@@ -108,10 +108,10 @@ printMixedStruct(mixedStruct *structToPrint)
 				   structToPrint->integer.max);
 			break;
 		case PGC_INT64:
-			printf("INT64\t%lld\t%lld\t%lld\t",
-				   (long long) structToPrint->integer8.reset_val,
-				   (long long) structToPrint->integer8.min,
-				   (long long) structToPrint->integer8.max);
+			printf("INT64\t%" PRId64 "\t%" PRId64 "\t%" PRId64 "\t",
+				   structToPrint->integer8.reset_val,
+				   structToPrint->integer8.min,
+				   structToPrint->integer8.max);
 			break;
 		case PGC_REAL:
 			printf("REAL\t%g\t%g\t%g\t",
